chore: rename weapp
